kernel-shark-qt: New way to visualize filtered events
The visualization of the entries filtered by event Id or by the
Advanced filtering is now different, compared to the visualization
of the entries, filtered by Process Id. While the entries, filtered
by Process Id are completely ignored (excluded from the graph),
the entries filtered by Event Id are taken into account when plotting
the area of each task, but do not show up as vertical ticks.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>