kernel-shark-qt: Add a plugin for renamed threads.
Sometimes is very handy to use pthread_setname_np() to differentiate
running tasks, for example when debugging an applications which is spawning
a lot of threads. This plugin is used to set the name of the task (shown by
KernelShark) to be the last name associated with it in the plotted trace.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>